+++ test/definition.js
@@ -47,6 +47,17 @@ describe('Implicit @define', function() {
       util.assertSingleFailure(css, {implicitComponents: ['css/c/*.scss'], preset: 'bem'}, null, filenameWithUnderscore);
     });
 
+    it('must use the parent directory if the file name is index.css', function() {
+      util.assertSuccess('.implicit-component {}', {
+        implicitComponents: true,
+        preset: 'bem',
+      }, null, process.cwd() + '/css/implicit-component/index.css');
+      util.assertSuccess('.implicit-component {}', {
+        implicitComponents: true,
+        preset: 'bem',
+      }, null, process.cwd() + '/css/implicit-component/_index.css');
+    });
+
     it('must complain about component name', function() {
       util.assertSingleFailure(
         css,